feat: 유효한 회원 정보 연결

See merge request cloud_track/class_04/web_project3/team01/backend!59

+++ src/main/java/com/team01/billage/user/domain/CustomUserDetails.java
@@ -17,8 +17,8 @@ public class CustomUserDetails implements UserDetails {
         this.jws = jws;
     }
 
-    public int getId() {
-        return jws.getBody().get("accountId", Integer.class);
+    public Long getId() {
+        return jws.getBody().get("accountId", Long.class);
     }
 
     public String getEmail() {
